requirements.txt
python-telegram-bot==21.6
aiofiles==23.2.1
aiohttp==3.9.5
pyte==0.8.2
pywinpty>=2.0.0; sys_platform == "win32"
mss>=9.0.0
windows-capture>=2.0.0; sys_platform == "win32"
Pillow>=10.0.0
pyautogui>=0.9.54
PyYAML>=6.0
croniter>=2.0.0

# FastMCP SDK — used by mcp_server/ and bridged into the proxy's
# managed-tool registry. Pip package "mcp"; import name "mcp".
mcp>=1.0.0

# Windows Job Object integration (proc_group.py) + elevated process
# introspection. Import name "win32api" etc.
pywin32>=306; sys_platform == "win32"

# System tray + settings window (Qt)
PySide6>=6.6
